+++ app/service/jarvis/execute/hash.rb
@@ -61,6 +61,9 @@ class Jarvis::Execute::Hash < Jarvis::Execute::Executor
         when :"logic.break"
           loop_exit = true
           break [eval_block(arg)] # still evaling here so that it increments
+        when :"task.exit"
+          loop_exit = true
+          break [eval_block(arg)] # still evaling here so that it increments
         else
           eval_block(arg).tap { |arg_val|
             break arg_val if jil.ctx.delete(:next)
